refactor: useEllipsis hook
